+++ knowthenics-data/src/main/java/ru/dpopkov/knowthenics/model/SourceType.java
@@ -1,0 +1,13 @@
+package ru.dpopkov.knowthenics.model;
+
+public class SourceType {
+    private String name;
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+}