+++ History.md
@@ -2,3 +2,4 @@
 
 #### 2020-12-09
 * Created multi-module project. Closes #1.
+* Created POJO Data Model. Closes #2.